Restaurant_app

+++ app/src/main/java/com/example/restaurant_app/model/Product.java
@@ -26,10 +26,10 @@ public class Product {
         private String description;
         @SerializedName("originalPrice")
         @Expose
-        private Integer originalPrice;
+        private Double originalPrice;
         @SerializedName("offerPrice")
         @Expose
-        private Integer offerPrice;
+        private Double offerPrice;
         @SerializedName("imageUrl")
         @Expose
         private String imageUrl;
@@ -91,23 +91,23 @@ public class Product {
             this.description = description;
         }
 
-        public Integer getOriginalPrice() {
-            return originalPrice;
-        }
+    public Double getOriginalPrice() {
+        return originalPrice;
+    }
 
-        public void setOriginalPrice(Integer originalPrice) {
-            this.originalPrice = originalPrice;
-        }
+    public void setOriginalPrice(Double originalPrice) {
+        this.originalPrice = originalPrice;
+    }
 
-        public Integer getOfferPrice() {
-            return offerPrice;
-        }
+    public Double getOfferPrice() {
+        return offerPrice;
+    }
 
-        public void setOfferPrice(Integer offerPrice) {
-            this.offerPrice = offerPrice;
-        }
+    public void setOfferPrice(Double offerPrice) {
+        this.offerPrice = offerPrice;
+    }
 
-        public String getImageUrl() {
+    public String getImageUrl() {
             return imageUrl;
         }
